# ShopAPI/ShopAPI/Models/OrderDto.cs
using System.ComponentModel.DataAnnotations;

namespace ShopAPI.Models
{
    public class OrderDto
    {
        [Required]
        public string ProductIdentifiers { get; set; } = "";

        [Required, MinLength(30), MaxLength(100)]
        public string DeliveryAddress { get; set; } = "";

        [Required]
        public string PaymentMethod { get; set; } = "";
    }
}