# ShopAPI/ShopAPI/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My Shop API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrderDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "paymentStatus",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "orderStatus",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "OrderDto": {
      "required": [
        "deliveryAddress",
        "paymentMethod",
        "productIdentifiers"
      ],
      "type": "object",
      "properties": {
        "productIdentifiers": {
          "minLength": 1,
          "type": "string"
        },
        "deliveryAddress": {
          "maxLength": 100,
          "minLength": 30,
          "type": "string"
        },
        "paymentMethod": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please enter token"
    }
  }
}